+++ common/src/main/java/io/github/flemmli97/runecraftory/common/entities/monster/boss/EntityAmbrosia.java
@@ -14,10 +14,8 @@ import io.github.flemmli97.tenshilib.api.entity.AnimationHandler;
 import io.github.flemmli97.tenshilib.common.utils.RayTraceUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.damagesource.DamageSource;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.MoverType;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
@@ -92,17 +90,20 @@ public class EntityAmbrosia extends BossMonster {
     }
 
     @Override
-    public void aiStep() {
-        if (this.getAnimationHandler().isCurrentAnim(angry.getID(), defeat.getID())) {
-            double g = -0.2;
-            if (this.isInWater())
-                g /= 16;
-            else if (this.isInLava())
-                g /= 8;
-            this.move(MoverType.SELF, new Vec3(0, g, 0));
+    protected boolean isImmobile() {
+        return super.isImmobile() || this.getAnimationHandler().isCurrentAnim(angry.getID(), defeat.getID());
+    }
+
+    @Override
+    public void push(double x, double y, double z) {
+        if (this.getAnimationHandler().isCurrentAnim(pollen.getID(), angry.getID(), defeat.getID()))
             return;
-        }
-        super.aiStep();
+        super.push(x, y, z);
+    }
+
+    @Override
+    public boolean shouldFreezeTravel() {
+        return this.getAnimationHandler().isCurrentAnim(wave.getID());
     }
 
     @Override
@@ -224,13 +225,6 @@ public class EntityAmbrosia extends BossMonster {
                 this.level.addFreshEntity(wave);
             }
         }
-    }
-
-    @Override
-    public void push(Entity entityIn) {
-        if (this.getAnimationHandler().isCurrentAnim(pollen.getID()))
-            return;
-        super.push(entityIn);
     }
 
     @Override